+++ lib/packages/wikibase/data-model/RELEASE-NOTES.md
@@ -11,6 +11,7 @@
 * Added `ItemNotFoundException`
 * Empty strings are now detected as invalid language codes in the term classes
 * `Item` constructor parameters for site links and statements are optional now
+* `Fingerprint` constructor can be empty
 
 ## Version 2.4.1 (2014-11-26)